Code compiles

+++ replacement_policy.py
@@ -7,10 +7,10 @@ class replacement_policy(object):
     '''
 
     def __init__(self, size : int = 0, associativity : int = 0) -> None:
-        self._cache_size          = size
-        self._cache_associativity = associativity
-        self._cache_entries       = self._cache_size / 4 # integer entries in the Cache are 4 bytes
-        self._cache_sets          = self._cache_entries / self._cache_associativity
+        self._cache_size          = int(size)
+        self._cache_associativity = int(associativity)
+        self._cache_entries       = int(self._cache_size / 4) # integer entries in the Cache are 4 bytes
+        self._cache_sets          = int(self._cache_entries / self._cache_associativity)
 
     def compute_to_evict(self) -> None :
         '''Runtime error if this function is invoked. This is meant to be
